Show kernel status (currently yellow/red) when opening a notebook
